requirements.txt
coloredlogs==15.0
Flask~=1.1.2
python-dotenv~=0.17.1
onnx==1.9.0
onnxruntime==1.7.0
torch==1.7.0
transformers==4.5.0
Flask-Cors==3.0.10
gunicorn==20.1.0